+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasetsop.py
@@ -96,7 +96,7 @@ class ListDatasetsData(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-11-05T05:24:55.173Z")
+    updated: Optional[datetime] = parse_datetime("2025-11-05T11:09:10.226Z")
     r"""The date and time the resource was last updated"""