+++ packages/orq-rc/src/orq_ai_sdk/models/updatebudgetop.py
@@ -215,5 +215,5 @@ class UpdateBudgetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-11-05T05:24:55.569Z")
+    updated: Optional[datetime] = parse_datetime("2025-11-05T11:09:10.517Z")
     r"""The date and time the resource was last updated"""